bug: none
Update to reflect change to "long" for ByteUtil.copy() return value.

+++ ClamScanner/src/java/com/zimbra/clam/ClamScanner.java
@@ -173,7 +173,7 @@ public class ClamScanner extends UploadScanner implements ZimbraExtension {
                 dataSocket.getOutputStream().write(data);
                 if (mLog.isDebugEnabled()) { mLog.debug("wrote " + data.length + " bytes"); }
             } else {
-                int count = ByteUtil.copy(is, false, dataSocket.getOutputStream(), false);
+                long count = ByteUtil.copy(is, false, dataSocket.getOutputStream(), false);
                 if (mLog.isDebugEnabled()) { mLog.debug("copied " + count + " bytes"); }
             }
             dataSocket.close();
